Reformatting form field descriptions.
A little more verbose, but I think generally more readable.
It means more lines are > 80 chars, but most of the modified lines
already were.

Should possibly inline the $args values into the #description t() calls.

Removed the incorrect use of t() with the substitutions for the
js_css_noscript description.